[📚 hlst_utils.c] Reorganized t_hlst functions
